moving AdminMenu
