+++ client/src/components/Edit.jsx
@@ -1,12 +1,7 @@
 import React from "react";
 
-
 function Edit() {
-  
-  return (
-  <form></form>);
+  return <form></form>;
 }
 
 export default Edit;
-
-
